% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grid_points}
\alias{grid_points}
\title{Get Grid of Continuous USA}
\usage{
grid_points(resolution = 0.1)
}
\arguments{
\item{resolution}{Resolution of graphic, default is 0.1}
}
\value{
dataframe with the following columns:
\describe{
    \item{Var1}{Longitude}
    \item{Var2}{Latitude}
    }
}
\description{
This function makes a map of the USA using grid points. All grid points
fall within the continuous USA
}
\examples{
#Make grid points with default resolution
grid_points()

}
