% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_interpolations}
\alias{plot_interpolations}
\title{Plot interpolation}
\usage{
plot_interpolations(
  df,
  col1,
  type = 1,
  col2 = NULL,
  Title = NULL,
  size_name = NULL
)
}
\arguments{
\item{df}{dataframe with following components
\describe{
    \item{LONGITUDE}{Longitude from grid_points}
    \item{LATITUDE}{Latitude from grid_points}
    \item{.}{Variable for color}
    }}

\item{col1}{Variable for color}

\item{type}{type of plot needed. 1 is for continuous color variables,
and 2 is for continuous variables the are very small. (Optional)}

\item{col2}{Variable for size (Optional)}

\item{Title}{Title of plot}

\item{size_name}{Title size on the legend}
}
\value{
A plot of spatial data
}
\description{
This function uses ggplot2 to make plots of spatial data
}
\examples{
#Type 1, default colors, no size
plot_interpolations(df = preds, col1 = preds$AVERAGE)

}
