% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{station_grid_points}
\alias{station_grid_points}
\title{Interpolates values given grid points}
\usage{
station_grid_points(df, grid_points, param = "T_DAILY_AVG")
}
\arguments{
\item{df}{dataframe with the following components
\describe{
    \item{LONGITUDE}{Longitude for training the model}
    \item{LATITUDE}{Latitude for training the model}
    \item{.}{A target variable of choice}
    }}

\item{grid_points}{dataframe with the following components:
\describe{
    \item{.}{Longitude for predictions}
    \item{.}{Latitude for predictions}
    }}

\item{param}{Optional parameter for the target variable}
}
\value{
a dataframe with the following components:
\describe{
    \item{LONGITUDE}{Longitude from grid_points}
    \item{LATITUDE}{Latitude from grid_points}
    \item{AVERAGE}{Predictions from the model}
    }
}
\description{
This function is predicts a target variable, given grid points using a
spatial model
}
\examples{
#Using default target
df <- weather_data[1:10000,]
grid <- grid_points(0.75)
station_grid_points(df,grid)

}
