% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Weather Stations}
\format{
a dataframe with 236 rows and 5 columns
\describe{
    \item{WBAN_ID}{The station WBAN number.}
    \item{STATE_PROVINCE}{State where weather station is located.}
    \item{STATION_NAME}{Name of weather station.}
    \item{LONGITUDE}{Station longitude, using WGS-84.}
    \item{LATITUDE}{Station latitude, using WGS-84.}
     }
}
\usage{
stations
}
\description{
A dataset with weather stations.
}
\keyword{datasets}
