% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{time_series}
\alias{time_series}
\title{A function for extracting the time series for a specific station by station id.}
\usage{
time_series(station_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{station_id}{WBAN ID of the weather station}

\item{start_date}{Start date of time series}

\item{end_date}{End date of time series}
}
\value{
a dataframe for a specific weather station. For information about
each column please refer to the weather_data documentation
}
\description{
Get a dataframe for a specific station with optional arguments regarding
the date.
}
\examples{
#Get data regarding station ID 3047
time_series(3047)

#Get data regarding station ID 3047 from 2003-05-22 to 2003-05-30
time_series(3047,"2003-05-22","2003-05-30")
}
