% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{yearly_cycle}
\alias{yearly_cycle}
\title{A function to estimate the yearly cycle for a single station}
\usage{
yearly_cycle(station_id)
}
\arguments{
\item{station_id}{WBAN ID of the weather station}
}
\value{
a dataframe with the following columns:
\describe{
    \item{day}{Day of the year (1-365)}
    \item{temp}{Estimate temperature for the day}
    }
}
\description{
Returns the estimated expected temperature for each day of the year.
Uses a second order regression model.
}
\examples{
#Get data regarding station ID 3047
yearly_cycle(3047)

}
