% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{yearly_trend}
\alias{yearly_trend}
\title{A function to estimate look at temperature trend over time per year.}
\usage{
yearly_trend(station_id, type = "T_DAILY_AVG")
}
\arguments{
\item{station_id}{WBAN ID for a specific station}

\item{type}{The temperature type}
}
\value{
a vector containing the trend estimate, p-value, and standard error.
}
\description{
Using a simple linear model to estimate trend.
}
\examples{
#Get data regarding station ID 3047
yearly_trend(3047)

}
