% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_interpolations}
\alias{plot_interpolations}
\title{Plot interpolation}
\usage{
plot_interpolations(
  df,
  col1,
  type = 1,
  col2 = NULL,
  Title = NULL,
  size_name = NULL,
  Big_Title = NULL
)
}
\arguments{
\item{df}{dataframe with following components
\describe{
    \item{LONGITUDE}{Longitude from grid_points}
    \item{LATITUDE}{Latitude from grid_points}
    \item{.}{Variable for color}
    }}

\item{col1}{Variable for color}

\item{type}{type of plot needed. 1 is for continuous color variables with no
background, and 2 is for continuous variables the are very small. Type 3
has a background(Optional)}

\item{col2}{Variable for size (Optional)}

\item{Title}{Title of color}

\item{size_name}{Title size on the legend}

\item{Big_Title}{Title of Plot}
}
\value{
A plot of spatial data
}
\description{
This function uses ggplot2 to make plots of spatial data
}
\examples{
df <- data.frame(station = c(),trend = c(),p_value = c(), se = c())
#Making a data frame for trend estimates of each station
for(station in stations){
new_df <- weather_data \%>\% filter(WBANNO == station)
LONGITUDE <- unique(new_df$LONGITUDE)
LATITUDE <- unique(new_df$LATITUDE)
df <- rbind(df,c(station,LONGITUDE,LATITUDE,
                 yearly_trend(station_id = station)))
}
colnames(df) <- c("station","LONGITUDE","LATITUDE","trend", "pvalue","se")
#Plotting trend estimates
plot_interpolations(df = df, col1 = df$trend, type = 3,col2 = df$pvalue,
Title = "Trend", size_name = "p_value", Big_Title = "Temperature Trend")

}
